fix: corrected a logical error in the calculation of cumulative yield per LP token

+++ programs/tyrbine-program/src/states/staker.rs
@@ -4,6 +4,6 @@ use anchor_lang::prelude::*;
 pub struct Staker {
     pub owner: Pubkey,
     pub vault: Pubkey,
-    pub last_cumulative_yield: u64,
+    pub last_cumulative_yield: u128,
     pub pending_claim: u64
 }
